restructure the api
