+++ app/__init__.py
@@ -2,9 +2,8 @@ from flask import Flask
 
 # routes
 from app.routes.main import main
-from app.routes.api import api
 from app.routes.auth import auth
-from app.routes.events import events
+from app.routes.api.v1 import events
 
 
 # db instance
@@ -22,8 +21,7 @@ def create_app(configuration):
 
 	# register blueprints
 	app.register_blueprint(main)
-	app.register_blueprint(api, url_prefix=app.config['API_BASE_URL'])
 	app.register_blueprint(auth, url_prefix='/auth')
-	app.register_blueprint(events, url_prefix='/events')
+	app.register_blueprint(events, url_prefix=app.config['API_BASE_URL'])
 	
 	return app
